# api.yaml
openapi: 3.0.3
info:
  title: Hospital Management System API
  description: >
    REST API for a hospital management system supporting three roles —
    Admin, Doctor, and Patient — with JWT-based authentication, appointment
    scheduling, treatment records, AI-powered patient summaries, and a
    post-appointment chat feature.
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Local development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Unauthorized

    # ─── Doctor ──────────────────────────────────────────────────────────────
    Doctor:
      type: object
      properties:
        doctor_id:
          type: integer
        doctor_username:
          type: string
        doctor_name:
          type: string
        doctor_email:
          type: string
          format: email
        doctor_type:
          type: string
        doctor_qualification:
          type: string
        doctor_experience:
          type: string
        doctor_description:
          type: string
        dep_id:
          type: integer
          nullable: true
...
    Patient:
      type: object
      properties:
        patient_id:
          type: integer
        patient_name:
          type: string
        patient_username:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        gender:
          type: string
        age:
          type: integer
        address:
          type: string
        is_blacklisted:
          type: boolean
    PatientRegisterRequest:
      type: object
      required:
        - name
        - username
        - email
        - password
      properties:
        name:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        address:
          type: string
        phone:
          type: string
        gender:
          type: string
        age:
          type: integer
...
    ChatMessage:
      type: object
      properties:
        message_id:
          type: integer
        appointment_id:
          type: integer
        sender_role:
          type: string
          enum: [DOCTOR, PATIENT]
        sender_id:
          type: integer
        message:
          type: string
        created_at:
          type: string
          format: date-time
    ChatMessageRequest:
      type: object
      required:
        - message
      properties:
        message:
          type: string

    # ─── Availability ────────────────────────────────────────────────────────
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string

# ═══════════════════════════════════════════════════════════════════════════════
# PATHS
# ═══════════════════════════════════════════════════════════════════════════════
paths:
...
  /api/auth/register/patient:
    post:
      tags: [Auth]
      summary: Register a new patient account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientRegisterRequest'
      responses:
        "201":
          description: Patient registered successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          patient:
                            $ref: '#/components/schemas/Patient'
        "400":
          description: Username or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login/patient:
    post:
      tags: [Auth]
      summary: Patient login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Login successful — refresh token set as HttpOnly cookie
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          patient:
                            $ref: '#/components/schemas/Patient'
                          access_token:
                            type: string
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Account is blacklisted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login/doctor:
    post:
      tags: [Auth]
      summary: Doctor login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Login successful — refresh token set as HttpOnly cookie
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          doctor:
                            $ref: '#/components/schemas/Doctor'
                          access_token:
                            type: string
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login/admin:
    post:
      tags: [Auth]
      summary: Admin login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Login successful — refresh token set as HttpOnly cookie
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user_type:
                            type: string
                            example: admin
                          access_token:
                            type: string
        "401":
          description: Invalid admin credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Return the currently authenticated user's profile and role
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          role:
                            type: string
                            enum: [admin, doctor, patient]
        "401":
          description: Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ────────────────────────────── ADMIN ───────────────────────────────────────
...
  /api/doctor/appointments/{app_id}/close-chat:
    post:
      tags: [Doctor]
      summary: Close the follow-up chat window for a completed appointment
      parameters:
        - in: path
          name: app_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chat closed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        "400":
          description: Appointment is not completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Doctor does not own this appointment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ────────────────────────────── PATIENT ─────────────────────────────────────
...
  /api/appointments/{app_id}/messages:
    parameters:
      - in: path
        name: app_id
        required: true
        schema:
          type: integer
    get:
      tags: [Chat]
      summary: >
        Retrieve all chat messages for a completed appointment.
        Returns a chat_open flag indicating whether new messages can still be sent
        (within 48 h of completion and not manually closed by the doctor).
      responses:
        "200":
          description: Chat messages and open/closed state
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          messages:
                            type: array
                            items:
                              $ref: '#/components/schemas/ChatMessage'
                          chat_open:
                            type: boolean
        "403":
          description: Appointment not completed, or user is not the doctor/patient for this appointment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Chat]
      summary: Send a message in a completed appointment's chat thread
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageRequest'
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ChatMessage'
        "400":
          description: Empty message body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Chat closed, 48-hour window expired, or unauthorized user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...